getRad: Download Radar Data for Biological Research

Load polar volume and vertical profile data for aeroecological research directly into R. With 'getRad' you can access data from several sources in Europe and the US and standardize it to facilitate further exploration in tools such as 'bioRad'.

See also

Author

Maintainer: Bart Kranstauber b.kranstauber@uva.nl (ORCID) (University of Amsterdam)

Authors:

Pieter Huybrechts pieter.huybrechts@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO))

Peter Desmet peter.desmet@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO))

Other contributors:

Cecilia Nilsson cecilia.nilsson@biol.lu.se (ORCID) (Lund University) [contributor]

Alexander Tedeschi at744@cornell.edu (ORCID) (Cornell Lab of Ornithology) [contributor]

Hidde Leijnse (ORCID) (Royal Netherlands Meteorological Institute) [contributor]

Bart Hoekstra (ORCID) (University of Amsterdam) [contributor]

University of Amsterdam (04dkp9463) [copyright holder]

Biodiversa+ (https://hirad.science/) [funder]