Retrieve vertical profile time series data from the Aloft data repository

Retrieve vertical profile time series data from the Aloft data repository

Usage

get_vpts(
  radar,
  date,
  source = c("baltrad", "uva", "ecog-04003"),
  return_type = c("vpts", "tibble")
)

Arguments

radar

The name of the radar (odim string) as a character string (e.g. "nlhrw" or "fikor").

date

Either a single date or a lubridate::interval

source

The source of the data. One of baltrad, uva or ecog-04003. Only one source can be queried at a time. If no source is provided baltrad is used.

return_type

The type of object that should be returned. By default the data is returned as a bioRad::summary.vpts object. If set to tibble, a dplyr::tibble() will be returned instead with an extra column for the radar source.

Value

By default, a vpts object is returned. See bioRad::summary.vpts for more information. When multiple radars are selected, a list of vpts objects will be returned instead. When return_type = "tibble", a single dplyr::tibble() is returned with an extra column for the radar source.

Details

By default, data from the aloft data repository is retrieved from https://aloftdata.s3-eu-west-1.amazonaws.com, this can be changed by setting options(getRad.aloft_data_url) to any desired url.

Examples

if (FALSE) { # interactive()

# Fetch vpts data for a single radar and date

get_vpts(radar = "bejab", date = "2023-01-01", source = "baltrad")

# Fetch vpts data for multiple radars and a single date

get_vpts(
  radar = c("dehnr", "deflg"),
  date = lubridate::ymd("20171015"),
  source = "baltrad"
)

# Fetch vpts data for a single radar and a date range

get_vpts(
  radar = "bejab",
  date = lubridate::interval(
    lubridate::ymd_hms("2023-01-01 00:00:00"),
    lubridate::ymd_hms("2023-01-02 00:14:00")
  ), source = "baltrad"
)

get_vpts("bejab", lubridate::interval("20210101", "20210301"), "bejab")

# Fetch vpts data for a single radar and a date range from a specific
# source

get_vpts(radar = "bejab", date = "2016-09-29", source = "ecog-04003")

# Return a tibble instead of a vpts object

get_vpts(
  radar = "chlem", date = "2023-03-10", source = "baltrad",
  as_tibble = TRUE
)
}